Security to all necessary endpoints were added

+++ src/comments/controllers/comments.controller.ts
@@ -1,35 +1,93 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Patch, Post, Req, UnauthorizedException } from '@nestjs/common';
 import { CommentsService } from '../services/comments.service';
 import { Comment } from '../entities/comment.entity';
+import { JwtService } from '@nestjs/jwt';
+import { Request } from 'express';
 
 @Controller('comments')
 export class CommentsController {
     constructor(
-        private commentsService: CommentsService
+        private commentsService: CommentsService,
+        private jwtService: JwtService
     ) {}
     
     @Get('all')
-    getAll(): Promise<Comment[]> {
-        return this.commentsService.getAll();
+    async getAll(@Req() request: Request): Promise<Comment[]> {
+        try {
+            const cookie = request.cookies['jwt']; // Request a cookie
+            const data = await this.jwtService.verifyAsync(cookie); // When the cookie is retrieved and still valid is going to return the user data (use_id and cookie's informations)
+    
+            if (!data) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.commentsService.getAll();
+        } catch (error) {
+            throw new UnauthorizedException();
+        }
     }
 
     @Get(':id')
-    getOne(@Param('id') id: number): Promise<Comment> {
-        return this.commentsService.getOneById(id);
+    async getOne(@Req() request: Request, @Param('id') id: number): Promise<Comment> {
+        try {
+            const cookie = request.cookies['jwt']; // Request a cookie
+            const data = await this.jwtService.verifyAsync(cookie); // When the cookie is retrieved and still valid is going to return the user data (use_id and cookie's informations)
+    
+            if (!data) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.commentsService.getOneById(id);
+        } catch (error) {
+            throw new UnauthorizedException();
+        }
     }
 
     @Post()
-    create(@Body() body: object): Promise<Comment> {
-        return this.commentsService.createComment(body);
+    async create(@Req() request: Request, @Body() body: object): Promise<Comment> {
+        try {
+            const cookie = request.cookies['jwt']; // Request a cookie
+            const data = await this.jwtService.verifyAsync(cookie); // When the cookie is retrieved and still valid is going to return the user data (use_id and cookie's informations)
+    
+            if (!data) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.commentsService.createComment(body);
+        } catch (error) {
+            throw new UnauthorizedException();
+        }
     }
 
     @Patch(':id')
-    update(@Param('id') id: number, @Body() body: object): Promise<Comment> {
-        return this.commentsService.updateComment(id, body);
+    async update(@Req() request: Request, @Param('id') id: number, @Body() body: object): Promise<Comment> {
+        try {
+            const cookie = request.cookies['jwt']; // Request a cookie
+            const data = await this.jwtService.verifyAsync(cookie); // When the cookie is retrieved and still valid is going to return the user data (use_id and cookie's informations)
+    
+            if (!data) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.commentsService.updateComment(id, body);
+        } catch (error) {
+            throw new UnauthorizedException();
+        }
     }
 
     @Delete(':id')
-    delete(@Param('id') id: number): Promise<Comment> {
-        return this.commentsService.deleteComment(id);
+    async delete(@Req() request: Request, @Param('id') id: number): Promise<Comment> {
+        try {
+            const cookie = request.cookies['jwt']; // Request a cookie
+            const data = await this.jwtService.verifyAsync(cookie); // When the cookie is retrieved and still valid is going to return the user data (use_id and cookie's informations)
+    
+            if (!data) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.commentsService.deleteComment(id);
+        } catch (error) {
+            throw new UnauthorizedException();
+        }
     }
 }
